command line syntax check - are all options (flags) are on the right postition inside the command line. rename color schemes and add a blue-yellow color scheme too (jsc#SLE-23727)
