+++ actions/actions_test.go
@@ -123,8 +123,9 @@ Daemon control:
   saptune daemon [ start | status | stop ]  ATTENTION: deprecated
   saptune service [ start | status | stop | restart | takeover | enable | disable | enablestart | disablestop ]
 Tune system according to SAP and SUSE notes:
-  saptune note [ list | verify | revertall | enabled | applied ]
-  saptune note [ apply | simulate | verify | customise | create | edit | revert | show | delete ] NoteID
+  saptune note [ list | revertall | enabled | applied ]
+  saptune note [ apply | simulate | customise | create | edit | revert | show | delete ] NoteID
+  saptune note verify [--colorscheme=<color scheme>] [--show-non-compliant] [NoteID]
   saptune note rename NoteID newNoteID
 Tune system for all notes applicable to your SAP solution:
   saptune solution [ list | verify | enabled | applied ]
@@ -246,8 +247,9 @@ Daemon control:
   saptune daemon [ start | status | stop ]  ATTENTION: deprecated
   saptune service [ start | status | stop | restart | takeover | enable | disable | enablestart | disablestop ]
 Tune system according to SAP and SUSE notes:
-  saptune note [ list | verify | revertall | enabled | applied ]
-  saptune note [ apply | simulate | verify | customise | create | edit | revert | show | delete ] NoteID
+  saptune note [ list | revertall | enabled | applied ]
+  saptune note [ apply | simulate | customise | create | edit | revert | show | delete ] NoteID
+  saptune note verify [--colorscheme=<color scheme>] [--show-non-compliant] [NoteID]
   saptune note rename NoteID newNoteID
 Tune system for all notes applicable to your SAP solution:
   saptune solution [ list | verify | enabled | applied ]
